+++ packages/api/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.10...@aws-amplify/api@5.0.11) (2023-01-19)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.9...@aws-amplify/api@5.0.10) (2023-01-13)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.9](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.8...@aws-amplify/api@5.0.9) (2023-01-10)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.7...@aws-amplify/api@5.0.8) (2022-12-27)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.6...@aws-amplify/api@5.0.7) (2022-12-16)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.5...@aws-amplify/api@5.0.6) (2022-12-15)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.5](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.4...@aws-amplify/api@5.0.5) (2022-12-06)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.3...@aws-amplify/api@5.0.4) (2022-11-23)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.2...@aws-amplify/api@5.0.3) (2022-11-19)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
+## [5.0.2](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.1...@aws-amplify/api@5.0.2) (2022-11-16)
+
+**Note:** Version bump only for package @aws-amplify/api
+
+
+
+
+
 ## [5.0.1](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api@5.0.0...@aws-amplify/api@5.0.1) (2022-11-11)
 
 **Note:** Version bump only for package @aws-amplify/api